chore: added tests for new pandoc functionalities
