+++ kreuzberg/_mime_types.py
@@ -54,24 +54,35 @@ IMAGE_MIME_TYPE_EXT_MAP: Final[Mapping[str, str]] = {
     "image/x-portable-pixmap": "ppm",
 }
 PANDOC_SUPPORTED_MIME_TYPES: Final[set[str]] = {
-    "application/csv",
-    "application/latex",
+    "application/csl+json",
+    "application/docbook+xml",
+    "application/epub+zip",
     "application/rtf",
     "application/vnd.oasis.opendocument.text",
     "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-    "application/x-csv",
+    "application/x-biblatex",
+    "application/x-bibtex",
+    "application/x-endnote+xml",
+    "application/x-fictionbook+xml",
+    "application/x-ipynb+json",
+    "application/x-jats+xml",
     "application/x-latex",
-    "application/x-rtf",
-    "application/x-vnd.oasis.opendocument.text",
+    "application/x-opml+xml",
+    "application/x-research-info-systems",
+    "application/x-typst",
     "text/csv",
-    "text/latex",
-    "text/rst",
-    "text/rtf",
     "text/tab-separated-values",
-    "text/x-csv",
-    "text/x-latex",
+    "text/troff",
+    "text/x-commonmark",
+    "text/x-dokuwiki",
+    "text/x-gfm",
+    "text/x-markdown",
+    "text/x-markdown-extra",
+    "text/x-mdoc",
+    "text/x-multimarkdown",
+    "text/x-org",
+    "text/x-pod",
     "text/x-rst",
-    "text/x-tsv",
 }
 
 SUPPORTED_MIME_TYPES: Final[set[str]] = (
